adding app file and writing MyApp with a sngleton pattern.
