+++ lib/main.dart
@@ -1,21 +1,4 @@
-
 import 'package:flutter/material.dart';
+import 'package:advanced_flutter_arabic/app/app.dart';
 
-void main() {
-  runApp(const MyApp());
-}
-
-
-class MyApp extends StatefulWidget {
-  const MyApp({super.key});
-
-  @override
-  State<MyApp> createState() => _MyAppState();
-}
-
-class _MyAppState extends State<MyApp> {
-  @override
-  Widget build(BuildContext context) {
-    return const Placeholder();
-  }
-}
+void main() => runApp(MyApp());
