--- lib/main.dart~
@@ -1,7 +1,0 @@
-
-import 'package:flutter/material.dart';
-
-void main() {
-  runApp(const MyApp());
-}
-
